refactor: 용어집 퀴즈 폴더 위치 수정
